requirements.txt
PyYAML==6.0
fastapi==0.75.0
uvicorn==0.17.6
filelock==3.6.0
